Fix args and created Response mixins
Fixes:

- Path args were not passed to endpoint handlers
- Respone Mixins were added to simplify making Response types
  that can define secure/insecure cookies and eventually more mixins.
